+++ include/malloy/particles/particles.hpp
@@ -1,6 +1,10 @@
 #pragma once
 
-// Umbrella header for malloy_particles: 2D colliding particles.
+// Umbrella header for malloy_particles: colliding particles in two dimensions
+// and (since M32) three.
 #include <malloy/particles/particle2d.hpp>
+#include <malloy/particles/particle3d.hpp>
+#include <malloy/particles/particle3d_settings.hpp>
+#include <malloy/particles/particle3d_world.hpp>
 #include <malloy/particles/particle_settings.hpp>
 #include <malloy/particles/particle_world.hpp>